fix(cluster): 优化集群创建事物
