+++ pkg/service/cluster_iaas.go
@@ -13,20 +13,23 @@ import (
 	"github.com/KubeOperator/KubeOperator/pkg/cloud_provider/client"
 	"github.com/KubeOperator/KubeOperator/pkg/constant"
 	"github.com/KubeOperator/KubeOperator/pkg/db"
+	"github.com/KubeOperator/KubeOperator/pkg/dto"
 	"github.com/KubeOperator/KubeOperator/pkg/model"
 	"github.com/KubeOperator/KubeOperator/pkg/model/common"
 	"github.com/KubeOperator/KubeOperator/pkg/repository"
 	"github.com/KubeOperator/KubeOperator/pkg/util/ipaddr"
 	"github.com/KubeOperator/KubeOperator/pkg/util/kotf"
 	"github.com/KubeOperator/KubeOperator/pkg/util/lang"
+	"github.com/jinzhu/gorm"
 )
 
-type ClusterIaasService interface {
-	Init(name string) error
+type ClusterCreateHelper interface {
+	LoadMetalNodes(creation *dto.ClusterCreate, cluster *model.Cluster, tx *gorm.DB) error
+	LoadPlanNodes(cluster *model.Cluster) error
 }
 
-func NewClusterIaasService() ClusterIaasService {
-	return &clusterIaasService{
+func NewClusterCreateHelper() ClusterCreateHelper {
+	return &clusterCreateHelper{
 		clusterRepo:         repository.NewClusterRepository(),
 		nodeRepo:            repository.NewClusterNodeRepository(),
 		hostRepo:            repository.NewHostRepository(),
@@ -36,7 +39,7 @@ func NewClusterIaasService() ClusterIaasService {
 	}
 }
 
-type clusterIaasService struct {
+type clusterCreateHelper struct {
 	clusterRepo         repository.ClusterRepository
 	hostRepo            repository.HostRepository
 	nodeRepo            repository.ClusterNodeRepository
@@ -45,43 +48,94 @@ type clusterIaasService struct {
 	vmConfigRepo        repository.VmConfigRepository
 }
 
-func (c clusterIaasService) Init(name string) error {
-	cluster, err := c.clusterRepo.Get(name)
-	if err != nil {
-		return err
+func (c clusterCreateHelper) LoadMetalNodes(creation *dto.ClusterCreate, cluster *model.Cluster, tx *gorm.DB) error {
+	workerNo, masterNo, firstMasterIP := 1, 1, ""
+	for _, nc := range creation.Nodes {
+		n := model.ClusterNode{
+			ClusterID: cluster.ID,
+			Role:      nc.Role,
+		}
+
+		var host model.Host
+		if err := tx.Set("gorm:query_option", "FOR UPDATE").Where("name = ?", nc.HostName).First(&host).Error; err != nil {
+			tx.Rollback()
+			return fmt.Errorf("can not find host %s cluster id ", nc.HostName)
+		}
+		host.ClusterID = cluster.ID
+		if err := tx.Save(&host).Error; err != nil {
+			tx.Rollback()
+			return fmt.Errorf("can not update host %s cluster id ", nc.HostName)
+		}
+
+		clusterResource := model.ClusterResource{
+			ClusterID:    cluster.ID,
+			ResourceID:   host.ID,
+			ResourceType: constant.ResourceHost,
+		}
+		if err := tx.Create(&clusterResource).Error; err != nil {
+			tx.Rollback()
+			return fmt.Errorf("can bind host %s to cluster", nc.HostName)
+		}
+
+		switch cluster.NodeNameRule {
+		case constant.NodeNameRuleDefault:
+			if n.Role == constant.NodeRoleNameMaster {
+				n.Name = fmt.Sprintf("%s-%s-%d", cluster.Name, constant.NodeRoleNameMaster, masterNo)
+				if len(firstMasterIP) == 0 {
+					firstMasterIP = n.Host.Ip
+				}
+				masterNo++
+			} else {
+				n.Name = fmt.Sprintf("%s-%s-%d", cluster.Name, constant.NodeRoleNameWorker, workerNo)
+				workerNo++
+			}
+		case constant.NodeNameRuleIP:
+			n.Name = host.Ip
+		case constant.NodeNameRuleHostName:
+			n.Name = host.Name
+		}
+
+		n.HostID = host.ID
+		if err := tx.Create(&n).Error; err != nil {
+			return fmt.Errorf("can not create  node %s reason %s", n.Name, err.Error())
+		}
+		n.Host = host
+		cluster.Nodes = append(cluster.Nodes, n)
 	}
-	if cluster.Provider == constant.ClusterProviderBareMetal || len(cluster.Nodes) > 0 {
+	if cluster.SpecConf.LbMode == constant.LbModeInternal {
+		cluster.SpecConf.LbKubeApiserverIp = firstMasterIP
+	}
+	cluster.SpecConf.KubeRouter = firstMasterIP
+	return nil
+}
+
+func (c clusterCreateHelper) LoadPlanNodes(cluster *model.Cluster) error {
+	if len(cluster.Nodes) > 0 {
 		return nil
 	}
-	plan, err := c.planRepo.GetById(cluster.PlanID)
+	tx := db.DB.Begin()
+	hosts, err := c.createHosts(*cluster, cluster.Plan)
 	if err != nil {
 		return err
 	}
-	hosts, err := c.createHosts(cluster, plan)
-	if err != nil {
-		return err
-	}
-	err = c.hostRepo.BatchSave(hosts)
-	if err != nil {
+	if err := tx.Create(&hosts).Error; err != nil {
+		tx.Rollback()
 		return err
 	}
 
-	k := kotf.NewTerraform(&kotf.Config{Cluster: name})
-	err = doInit(k, plan, hosts)
-	if err != nil {
-		for i := range hosts {
-			hosts[i].ClusterID = ""
-			_ = db.DB.Delete(&hosts[i])
-		}
-		return err
-	}
-	if err := c.hostRepo.BatchSave(hosts); err != nil {
+	k := kotf.NewTerraform(&kotf.Config{Cluster: cluster.Name})
+	if err = doInit(k, cluster.Plan, hosts); err != nil {
+		tx.Rollback()
 		return err
 	}
 
-	var projectResources []model.ProjectResource
+	var (
+		projectResources []model.ProjectResource
+		clusterResources []model.ClusterResource
+	)
 	prs, err := c.projectResourceRepo.ListByResourceIDAndType(cluster.ID, constant.ResourceCluster)
 	if err != nil {
+		tx.Rollback()
 		return err
 	}
 	for _, host := range hosts {
@@ -90,31 +144,34 @@ func (c clusterIaasService) Init(name string) error {
 			ResourceID:   host.ID,
 			ResourceType: constant.ResourceHost,
 		})
-		clusterResource := model.ClusterResource{
+		clusterResources = append(clusterResources, model.ClusterResource{
 			ClusterID:    cluster.ID,
 			ResourceID:   host.ID,
 			ResourceType: constant.ResourceHost,
-		}
-		if err := db.DB.Create(&clusterResource).Error; err != nil {
-			return err
-		}
+		})
 	}
-	err = c.projectResourceRepo.Batch(constant.BatchOperationCreate, projectResources)
-	if err != nil {
+	if err := tx.Create(&projectResources).Error; err != nil {
+		tx.Rollback()
+		return err
+	}
+	if err := tx.Create(&clusterResources).Error; err != nil {
+		tx.Rollback()
 		return err
 	}
 
-	nodes, err := c.createNodes(cluster, hosts)
+	nodes, err := c.createNodes(*cluster, hosts)
 	if err != nil {
+		tx.Rollback()
 		return err
 	}
-	if err := c.nodeRepo.BatchSave(nodes); err != nil {
+	if err := tx.Create(&nodes).Error; err != nil {
+		tx.Rollback()
 		return err
 	}
 	return nil
 }
 
-func (c clusterIaasService) createNodes(cluster model.Cluster, hosts []*model.Host) ([]*model.ClusterNode, error) {
+func (c clusterCreateHelper) createNodes(cluster model.Cluster, hosts []*model.Host) ([]*model.ClusterNode, error) {
 	masterNum := 0
 	workerNum := 0
 	var nodes []*model.ClusterNode
@@ -146,7 +203,7 @@ func (c clusterIaasService) createNodes(cluster model.Cluster, hosts []*model.Ho
 	return nodes, nil
 }
 
-func (c clusterIaasService) createHosts(cluster model.Cluster, plan model.Plan) ([]*model.Host, error) {
+func (c clusterCreateHelper) createHosts(cluster model.Cluster, plan model.Plan) ([]*model.Host, error) {
 	var hosts []*model.Host
 	masterAmount := 1
 	if plan.DeployTemplate != constant.SINGLE {
